// backend/src/domain/admission/enums/AdmissionErrorType.ts
export enum AdmissionErrorType {
    InvalidApplicationId = "Invalid application ID!",
    InvalidRegisterId = "Invalid register ID!",
    InvalidPaymentId = "Invalid payment ID!",
    ApplicationNotFound = "Application draft not found!",
    UnauthorizedAccess = "Unauthorized access to application!",
    InvalidSection = "Invalid section!",
    PaymentProcessingFailed = "Payment processing failed!",
    PaymentNotFound = "Payment not found!",
    PaymentMismatch = "Payment does not belong to this application!",
  }